Create live-static.yml
Updating all LU repos to the new actions-only Pages workflow in GH.  Taking a minimal first step, which is leveraging the existing build work (An existing action that deploys the static jekyll site to live (in gh-pages branch), dev (pushes static html to dev repo), test (same for test).  The new actions publish the generated, static html sites to the new Pages architecture. Future improvements would be to edit the existing workflow build tools to do this as a single step.